Add cluster autoscaler to EKS cluster
The autoscaler manages the autoscaling group for the worker nodes.

It automatically adjusts the size of the Kubernetes cluster when
there are pods that can't run due to insufficient resources or
there are nodes that are underutilised for a long time and we can
move their pods onto other nodes.

The implementation follows the AWS documentation:
* https://docs.aws.amazon.com/eks/latest/userguide/cluster-autoscaler.html
* https://github.com/terraform-aws-modules/terraform-aws-eks/blob/e3216e3cf80cb59089ba0e0365c6650520000aaf/docs/autoscaling.md

Trello: https://trello.com/c/pSUJPwuB/601-enable-cluster-autoscaler