Fix AWS tags for cluster nodes and node pool ASG.
This is a workaround for a couple of outstanding issues with the AWS
Terraform provider and the `eks` module, which were preventing default
tags from being propagated to the managed node group autoscale group.
Hopefully we'll be able to ditch the workaround at some point in future.
See the inline TODOs for details.

Also move root_volume_type into workers_group_defaults, since it makes
more sense there. (If we were to add another node group we'd want it to
default to gp3 storage and not the legacy, slightly pricier gp2 storage.)