Build govuk-terraform Image
We create our own image of terraform with aws-cli and jq since these 3
binaries are now needed to run terraform with assume role, see #378.

This image is based on what the PaaS team is doing
[here](https://github.com/alphagov/paas-docker-cloudfoundry-tools)

While in the trello card, it mentions to user ARG, it is not clear
whether we want to pass in these args from Concourse or not.  PaaS
directly modifies the Dockerfile.

We use a Debian image since awscli v2 is not easily compiled for
alpine base image, see
[issue](https://github.com/aws/aws-cli/issues/4685#issuecomment-615872019)

**Testing**
I modified the `build-images` pipeline and checked that it
is running
[successfully](https://cd.gds-reliability.engineering/teams/govuk-ci/pipelines/build-images/jobs/govuk-terraform/builds/9)

I also tested the images with a test `eks-test` pipeline
[here](https://cd.gds-reliability.engineering/teams/govuk-test/pipelines/eks-test)

Ref: 1.
[trello-card](https://trello.com/c/r2cYPT8p/621-create-a-tools-image-with-terraformawsclijq-for-concourse-to-replace-the-3rd-party-one-were-using-temporarily)