Add Grafana and Prometheus (MVP version)
This installs two helm charts (prometheus and grafana)
into a new 'monitoring' namespace.

This also configures grafana's elasticsearch and prom
data sources.

I've not defined dashboards in cluster-services since these
should be managed by users of cluster-services ideally not
in Terraform.

This is an MVP so I've not set up Grafana OAuth or
integrations with CloudWatch or DNS records for
Grafana/Prometheus/AlertManager. There are quite
a few other parts to add, but this should serve
our purposes for now.