Update EKS module from 17.1 to 17.8.
Changelog: https://github.com/terraform-aws-modules/terraform-aws-eks/blob/master/CHANGELOG.md#v1780---2021-09-03

There is one notable change here, which is that ebs_optimized is now
enabled by default (for instance types which support it) in the managed
node group launch template. See PR 1459 in
https://github.com/terraform-aws-modules/terraform-aws-eks/ (not linking
directly because I want to be a good neighbour and not have GH auto-spam
the PR with references).