Ensure unique names for IRSA IAM roles within the account.
Also construct the role names in the same way for all of the cluster
addons. I've chosen to include the cluster name in the IAM role names to
satisfy the uniqueness constraint but omit it from the k8s ClusterRole
names so that the ClusterRole names stay the same (and so that they
don't contain a spurious "govuk-" prefix).