Add a var for passing force_destroy to resource in non-prod.
This is needed for the Route53 zone which will have records added to it
by the external-dns addon, so that we can still run `terraform destroy`
in the test account without having to manually go and delete leftover
DNS records.

Defaults to false so that it'll be set safely in prod even if forgotten.

Considered using a conditional on govuk_environment, but decided that'd
be an unnecessary abstraction leak as well as putting unnecessary logic
in the config. This way is more explicit and it's easier to see what's
going on.