Use iam-assumable-role module in aws_lb_controller_iam.tf
We're using this for the other cluster addons and it reduces boilerplate
(albeit in return for a bit of extra complexity), so let's be
consistent.

Also, put aws-lb-controller in the cluster-services namespace instead of
kube-system. (There was a copy-pasto last time where the namespace in
the assume-role policy didn't match the namespace of the service account.)