Add pipeline to build govuk-ruby base images
We create a new pipeline to build the govuk-ruby base images on
which we will build the govuk apps images.
We need to build a base image for each of the version of ruby that
are used by the govuk apps.
The images are built daily and pushed to ECR only if some packages
have been updated compared to latest.