Add Publisher Helm Chart
This PR adds the Publisher Helm Chart.

Compared to the previous apps, publisher works in 2 modes:
web and worker. This is accounted for in the chart.

We also include a Kubernetes job that does the db migration if enabled
in the values.yaml file on each loading of the chart in Kubernetes.

**Notes**
Rolling of deployments when contents of secrets change can be done via
tools such as [stakater/reloader](https://github.com/stakater/Reloader)
as discusss.

Ref:
1. [trello card](https://trello.com/c/bnJygqQl/640-add-database-migration-job-to-helm-charts)