Fix creation of TLS cert for external_dns domain on initial turnup.
Turns out I broke turnup-from-scratch in efbe2cf. Terraform itself and
the AWS provider have some unfortunate limitations that prevent
filtering aws_acm_certificate.cluster_public.domain_validation_options
before passing it to resource for_each. (See the TODO in the this commit.)

Unfortunately this means that for now we'll have to put up with a
nonsense validation record in the ecs.$env.govuk.digital domain. That'll be
fixed when we get around to delegating the
{test,integration,staging}.publishing.service.gov.uk domains to zones in
their respective AWS accounts instead of hosting the whole lot in the
prod zone. That's probably somewhat entangled with sorting out
alphagov/govuk-dns though :/