Have Prometheus Operator read PrometheusRules from all namespaces.
This allows apps to define their own PrometheusRules etc. while still
allowing copies of apps to be deployed in extra namespaces for testing,
including their monitoring rules.

This might change if we were to decide we wanted more than one
Prometheus stack in the cluster, but I think that's highly unlikely: if
an application warrants having its metrics segregated on its own
Prometheus stack then it also shouldn't be sharing fate or security
boundaries with anything else in the cluster, so therefore should have
its own cluster. (Repeat after me: k8s is not designed for (hard)
multitenancy ;-)

There seem to be some issues with configuring the Prometheus Operator to
just look at all namespaces - there are conflicting docs and it doesn't
appear to be possible currently when using its Helm chart - so instead
we're configuring it to look at all namespaces except those labelled
"no_monitor". That's arguably better anyway as it allows for someone to
run their own Prometheus with prom operator if they really wanted to.