Include "redis" in the DNS name of the Redis cluster.
Looks like nothing much is using this yet so it's a good time to rename.

Also add a ".eks" into the name so that it's clear that this relates to
the EKS project. This is just a part of the name rather than a separate
zone, so it's easy enough to get rid of the ".eks" part in future (if
this Redis sticks around).

We can't just set shared_redis_cluster_name to "shared-redis.eks"
because Redis doesn't allow dots in replica names, hence we're appending
it at the point where we specify the Route53 CNAME record.

Trello: https://trello.com/c/Rp2xjAYJ/674