Simplify tfvars for the ECR module.
Instead of declaring and passing a variable for every AWS account that's
authorised to pull from the registry, just pass a list of principals.
It's simpler, more flexible and easier to understand and maintain.

Move the vars (actually now just one var) for the ECR module into
separate files to make it clear that they're specific to that module.
Helps to avoid warnings about unused vars when applying other modules.
Remove the now-disused vars from common.tfvars.

Add the staging account to the puller ACLs for the prod registry, since
it was missing.

Remove the prod account from the puller ACLs for the test registry,
since we don't ever want prod or staging to pull from the test registry.

Some trivial cleanup in ecr/main.tf and add a couple of TODOs.

Update the docs to match.