Add Powerusers as new category of k8s users
A new category of K8s users `powerusers` are added to the GOV.UK
cluster. These users will have admin access to a restricted
number of namespaces, `apps` namespace initially where GOV.UK
apps are deployed. In addition, the users will have read-access
at a cluster level.

A cluster role is used as it is possible to bind it to a set of roles
in different namespaces rather than a single namespace.

Ref:
1. [trello card](https://trello.com/c/IM6wt4Vj/718-create-basic-rbac-structure)
2. [k8s docs](https://kubernetes.io/docs/reference/access-authn-authz/rbac/#rolebinding-example)