Enables Argo-workflow Server UI
Enables the Argo-workflow server UI in the cluster. This will
allow users to view and potentially troubleshoot workflows.

Note that this work is about making the UI work but not make
the Single Sign-On (SSO) login feature works, which is detailed
in another [card](https://trello.com/c/hF9uauRU/587-google-sso-for-argo)

The URL for the service is:
https://argo-workflow.eks.<environment>.govuk.digital/

To get the authentication token:
1. install the `argo` cli using brew: `brew install argo`
   Must be version 3.2.5 or greater due to
   [bug](https://github.com/argoproj/argo-workflows/issues/7175)
2. get token: `gds aws govuk-<environment>-admin -- argo auth token`
3. copy whole token, starting with `Bearer ` and paste it in website.

Ref:
1. [trello card](https://trello.com/c/GHtiRMqe/781-argo-workflows-ui)