Manage GitHub resources with Terraform
This adds a new GitHub deployment which enables us to manage
GitHub resources with Terraform.

Currently it simply permits alphagov repositories with the tag
'govuk' to push images to ECR, by providing access to credentials.

There are various additional features we can add to simplify and automate
the configuration of new repositories and deploying new applications to
the Kubernetes cluster. This should tie in nicely with the ECR module,
though that module could also use the github provider to fetch the repositories
to create ECR registries, which would mean we don't need to manually add
to a list of applications in the module when we add a new app.