Update RabbitMQ broker version to what we're already running.
It's currently not possible to apply the govuk-publishing-infrastructure module
without keeping this in sync, because of a known bug in the AWS TF
provider: "Using auto_minor_version_upgrade (which we want) causes engine_version
drift." See [terraform-provider-aws
issue](https://github.com/hashicorp/terraform-provider-aws/issues) 14705
(not linking directly in order to avoid GitHub refspam).

We should probably just delete this MQ broker for the time being. We
only added it because the old MQ broke in the test environment; it's not
needed in integration and it probably won't be needed until after we've
switched the publishing apps to k8s (which is several milestones down
the road).