Add reusable GitHub Action workflow for deploying to EKS
This workflow will be run by actions on app repos after images
are built and pushed to ECR.

The workflow pulls alphagov/govuk-helm-charts, modifies a file
in charts/argocd-apps/app-images/integration, commits, and creates
a PR, and auto-merges the PR once the status checks pass.

It uses the ci-user credentials to achieve this, and we must give
the ci-user permission to bypass the code reviewer restriction.

This is a first pass at the implementation. Sean is looking at
using an Argo Workflow to achieve this, which might be nice since
we'll need that for release promotion beyond integration in future.