Attach search-api s3 bucket policies to EKS nodes
For search-api to run in EKS, it needs to access 2 S3 buckets.
We export the arn of the iam_policies of these 2 buckets in
[govuk-aws](https://github.com/alphagov/govuk-aws/pull/1558) to
the terraform remote state.

We then attach these iam_policies arns to attach the EKS nodes for
node. Later, we need to attach them to a custom  service account
used to run the search-api pod/container.