Fix dependency issues that were breaking cluster turnup.
Create the namespace for the ArgoCD-managed apps. While it's generally a
bad idea to create Kubernetes resources directly from Terraform, in this
case it's the lesser of evils because it's the simplest way to avoid a
circular dependency. (The alternative would be a Helm chart, but that
would just add complexity for no real benefit in this case.)

Set create_namespace on some helm_release resources that were missing
it.

Wait for aws_lb_controller to come up before installing Dex and ArgoCD
(which have Ingresses). In theory this isn't strictly necessary as
everything should converge eventually, but in practice it takes ages
because of exponential backoff and the turnup is slow enough as it is.
Slow things are a pain to test.