Update Argo helm charts
The main change is that ArgoCD helm chart now includes
ArgoCD Notifications so we remove the latter Helm chart.

ArgoCD now uses an official release rather than a candidate
release because the bug fix is now in the mainline.cc

Other minor changes are updating Argo Events and Argo
Workflows Helm charts to the latest version.