Add gitRef param to CI GitHub Actions workflows
This allows the two workflow involved in CI (image build and trigger
deploy) to accept a gitRef input to specify which git commit, tag or
branch to deploy. If a branch name is provided the latest commit on that
branch is deployed. Not further commits to that branch are not
automatically deployed.