Fix tagging images "latest" on older versions
This adds a check to the image build GitHub workflow to only tag images
as "latest" when the code is the latest commit on the default branch in
GitHub. e.g. the latest commit on main. This prevents images with older
versions of code being tagged as latest, for example building an image
for a commit on a branch. Previously the most recently push image was
tagged with "latest".