Add back the version numbers to helm_release resources.
Turns out that the Helm TF provider docs are misleading (though
technically correct); omitting the version attribute installs the latest
available version, but only once. On subsequent runs of `terraform
apply`, newer versions of the chart are ignored. This is worse than
having to update the version explicitly, so let's put the version
numbers back.

There appears to be no way (currently) to make TF do what we want, which
is to automatically keep up to date with the latest version of the chart
(for charts that we release ourselves) without needing to manually
`taint` the resource before running `terraform apply`.

Basically reverts e98d308.

Thanks to @samsimpson1 for finding this issue.