Skip dex approval screen to share data with application
This is to save some clicking when logging into an application. At the
moment every new session shows a screen that ask for your approval for
dex to share data with the application (argo, grafana etc.). Approval
for sharing data from connected IdP to Dex is separate process on IdP.