Publish Rubygem GitHub action workflow
This adds a GitHub action workflow that can be used to push a gem to
Rubygems. The motivation for adding this is that we want to reduce the
duplication of bash code used in mutltiple repositories before expanding
our usage of GitHub Actions to more gems.

An example of this action being used with a repo is available at: https://github.com/alphagov/rubocop-govuk/commit/1ac785c96aec754afe039c9b746bf6cee3b00f50

I felt that it made sense for this action to be included in the
govuk-infrastructure repo as we already have precedence for similar
actions, such as publishing to ECR, available.