Avoid creating an unused secondary SG for the cluster.
By default, the terraform-aws-modules/eks module creates an additional
security group for the cluster and adds some rules to it that refer to
the default node group SG that it creates, which we're also not using.

In order to avoid creating these unused (and therefore confusing) SGs,
we need to disable both, otherwise the module fails to plan.