add ability to specify dockerfile path
In most scenarios, the Dockerfile is found at the top-level directory
of the repo where the GitHub workflow is run but in some situations
such as signon-resources, the Dockerfile is found in a specific
directory.

This change allows the user to specify which path the Dockerfile is located.
Note that the context of the docker build is still the top level directory
of the repo.

This change is backward compatible with repos where Dockerfile is at top level.