Add ability to specify an app name to deploy job
The deploy job uses the repo-name as the app to be deployed.
This works for most scenarios where there is one app per repo.

We extend the deploy job so that it can optionally take a custom
app name to be deployed. This is required for apps such as
signon-resources where the app is in a shared repo.