Add test-rails GitHub Action reusable workflow
This adds an initial re-usuable workflow definition to be used by GOV.UK
Rails application to run CI tests. This is apart of work to migrate the
CI process from Jenkins to GitHub Actions. The workflow only supports
tests the require JS dependencies and MongoDB - as we are currently
focused on only migrating frontend app to being with. At time of writing
GitHub Actions does not support conditional `services` or using
`services` in workflows that call reusuable workflows, hence MongoDB is
created as a conditional step that can be enable with an workflow input.