Only publish gem if it is HEAD of default branch
This has been opened to resolve a scenario where errors can occur with
this action. These errors occur is if a build of the main branch is delayed
and this action runs after a new version of the gem has been created. This
happened with [1] where a delayed build of the main branch for gem version
6.8.2 ran after the release of 6.8.3.

I decided that the most effective way to resolve this issue was to only
run this action when the commit that is being built represents the HEAD
of the default branch. This seemed the only way that we could make the
situation of multiple merges and out-of-order builds fully predictable
as it considers repo state at the time of build.

This change does come with some situations that could be surprising: if
you merge two new releases rapidly, only the most recently merged one
could be deployed; if you merged a release and then a change that you
didn't want released it could be released. However what this does bring
is predictability for those niche situations. Without this change
multiple merges are unpredictable for gem releasing.

As far as I understand calling `git ls-remote origin HEAD` will always
get us the HEAD of the default branch and doesn't have to actually know
the name of that branch. Though I struggled to find documentation that
authoritatively stated this, just Stack Overflow [2].

---

In order to resolve this I did some research (and was given some
pointers) on approaches. This seems an appropriate place to document
them.

The interesting thing to consider, for a tool that does a form of
continuous deployment based off GitHub Actions, is that it isn't
guaranteed that a GitHub Action running against the main branch is
building against the HEAD of that branch. As we found in the situation
explained above, it's wholly possible to have HEAD run and then HEAD~1
run. This gets tricky if we want to deploy the code based off a
successful branch run as a later branch could mean we're actually
deploying a regression of the code that is commits before the main
branch's tip.

An idea on how this can be resolved is using the `concurrency` [3]
feature of GitHub Actions. This feature allows for there to only be one
job running at a time and to cancel any pending jobs when a new one is
triggered. I decided that using this actually risked making the
situation worse. For example in the above scenario of HEAD~1 being
kicked off after HEAD, assuming we had concurrency set at a branch
level, this would mean that HEAD could never get build as the queueing
of HEAD~1 would cancel it.

[1]: https://github.com/alphagov/govspeak/runs/6982747734?check_suite_focus=true
[2]: https://stackoverflow.com/questions/65362985/how-to-git-fetch-the-default-branch
[3]: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#concurrency