Add support to disable Brakeman for CI
This adds a new boolean input `runBrakeman` that can be used to disable
the Brakeman step. This is needed to re-use this workflow with
applications such as search-api which isn't a Rails app, hence doesn't
have support to be analysed by Brakeman. The CI workflow should be
renamed more generally, as it can be used to test non-rails apps.