Add AWS EFS CSI driver and create Clamav EFS
We want to create a EFS volume to store the Clamav virus
database. Before, we attempted to create the EFS in AWS only
and mount it as NFS in the Clamav pod but this was unsuccessful
since the EFS has root ownership and the pod is running as non-root.

We could create the access point of the EFS and sets the same ownership
as the pod in AWS but we would have to use the volume handler option of
k8s PersistentVolume for k8s to mount the access point properly. This
option requires the AWS EFS CSI driver as a prerequisite.

The chosen solution is:
1. install AWS EFS CSI driver and associated permissions
2. The storage class requires a EFS id so we create the EFS and
   pass it to the driver.