Increase startup wait time for MySQL in CI
On occasion 10s has not been long enough for MySQL to start and had
cause subsequent connection failure issue, ultimately causing CI to
fail. This is a quick fix to increase wait time to reduced the
flaky-ness of CI. The long term fix add a docker healthcheck and poll
until the check is health to continue.