Only build non-existing ECR images
If an ECR image exists, we do not build it again.

Future work:
Right now we use the tag to determine whether to rebuild or not and the tag comes from
GitHub repo. This logic may need to be changed when images are rebuild due to base image
changes due to updates.

Ref:
1. [trello card](https://trello.com/c/nFlcYHdt/1005-dont-re-build-image-if-already-existing)
2. [stack overflow](https://stackoverflow.com/questions/30543409/how-to-check-if-a-docker-image-with-a-specific-tag-exist-locally)