Add resuable GHA workflows for CI
This adds workflows that can be used by CI workflows in apps to run the
different process required. We are using reusable workflow to help
centralise and minimise the amount of config we have to manage. This
adds the initial set of jobs that are need by our frontend apps for CI.

The cache dependencies workflow can be run before other jobs to cache
shared dependencies.