Use build-and-push-image in build-github-cli
ci-ecr was removed in #769 in favour of build-and-push-image.yaml which was
added in #763. This workflow was missed in the update.

It looks like the github-cli image has since disappeared from the registry -
not sure if that's a problem with the image retention policy, or whether it was
accidentally deleted some other way...