Fix setting of image tags in docker build
This fixes a bug where only the first image tag would be pushed to ECR, not all
of them. The cause was an inaccuracy in bash syntax (that would have worked
fine in zsh):

In bash `echo "${array}"` only prints the first item in the array, not the
whole thing (note, this is different in zshell).

This meant that `echo "fullImageTagList=${FULL_IMAGE_TAG_LIST}" >> $GITHUB_OUTPUT`
was only collecting the first tag, not the full tag list.