Clarify setting of docker tags
The combination of

```
FULL_IMAGE_TAG_LIST=(`echo ${FULL_IMAGE_TAG_LIST}`)
```

(to turn the space-separated variable into an array)

and

```
${FULL_IMAGE_TAG_LIST[@]/#/-t }
```

(to expand the array into `-t first -t second` etc.)

...is very clever. To the point that it was beyond my bash knowledge to
interpret it without looking it up. It's a bit more verbose to use a loop, but
it's much clearer what's going on.