Upgrade Grafana DB from EoL Aurora PG10 to PG11.
RDS Aurora Postgres 10 is end-of-life in Jan 2023, so upgrade to 11.

This is still the Aurora Serverless v1 product for now - we can update
to v2 once this is out of the way. Or it might even be less hassle to
switch to an in-cluster database, seeing as how surprisingly toily it is
to change Aurora versions using Terraform.