Add precompile rails assets composite action
This is a composite GitHub Action that precompiles Rails assets and
caches them. The action caches the sprockets caches and any compiled
assets. The action will restore earlier caches for a git ref (i.e.
branch) in case of a cache miss. This may useful for speeding up asset
precompilation, as assets may have not changed or the sprockets caches
can be resused for partially changed assets.