Change the lint js workflow specific to Standardx
This makes the workflow more specific and clear in what its puporse is.
We currently use Standardx to lint all our JavaScript across all our
repositories. This also adds a input to change which files to lint, as
the varies across our repos, but the majority use the default setting.