Add workflows with yml extension
This is adds copies of workflows that previously had the extension
`.yaml`, with the new `.yml` extension. The old workflows have a
deprecation noticed and should not be updated. This is intermediate step
whilst other repositories that use these workflows are updated, then the
old workflows with `.yaml` will be removed.