Use fewer, bigger nodes.
Staging didn't fit in the previous maximum of 15 m6i.2xlarge nodes.
Rather than just increasing the maximum number of nodes, instead we
double the size of the nodes and adjust the limit to be 150% of what
we're currently using.

With this change, staging fits in 8 m6i.4xlarge nodes with less waste.

Also remove m5 instance types from the list of alternatives and replace
them with m6a, since m6a is much closer in performance characteristics.
We're not currently using spot instances anyway, but we might do on
integration or staging at some point.