Set allowed version of NATS Jetstream image
When starting the eventbus, its tries to use a specific version of the
NATS Jetstream image. The default config only allows use of the image
tagged latest, and hence the eventbus fails to start. Explicitly
allowing this version for now as a quick fix, until we figure out how to
not specify a specific allowed version.