Only query list of GitHub repositories once
This reduces the number of API calls to GitHub to get details for each
of the govuk repos. Instead querying GitHub for two sets of Terraform GitHub
repository resources, we query for one set and then generate the second
set by filtering. This is possible becuase the deployable repos are a
subset of all the repos that we need to mirror.