Use plain YAML for filebeat config.
This is so much easier to read and write than the HCL representation,
which has been making my head hurt the last couple of days.

We don't even need to refer to any HCL variables in this config file, so
it turns out to be trivial to switch to using its native representation
and still keep the ahead-of-time syntax checking.

The `yamlencode(yamldecode())` is not a no-op; we want the side effect
of it syntax-checking within Terraform at plan-time.

Tested: zero-diff in staging.