Fix filtering of healthcheck log lines.
The decoded JSON fields don't appear to be visible to the condition in
drop_event (whether the drop_event is defined for the container input or
globally), so let's just use `exclude_lines` on the raw input instead.

Also just exclude `/readyz` (kubelet -> nginx healthchecks) and not
`/healthcheck/live` since the latter is a much smaller proportion of log
lines and surprisingly there's some interesting weirdness going on with
at least one Ruby app's handling of those 🙃

Tested: applied in staging; healthcheck messages are gone and other logs
are still flowing.