Stop scraping kube-apiserver.
Turns out scraping the apiservers is surprisingly expensive and we're not
even using any of those metrics, so let's ditch it. This should make the
monitoring we do care about a bit more robust. We don't run the
apiservers anyway; that's in Amazon-land and there are metrics in
CloudWatch if we ever needed to look.