Drop some more not-useful log fields.
These fields have identical values in every log message from the last 7
days in production, are highly unlikely to change in the foreseeable
future and unlikely to be of interest even if they do change.
`instance-type` is an exception to this, but that one's not worth
logging because we already have `node.labels.kubernetes_io/arch` for
when we start moving to arm64 nodes.