Add Tempo helm chart deployment
This adds Terraform to deploy the Tempo helm chart. We've enabled
persistent disk for the ingesters to prevent data loss if pods a
terminated. We've also enabled remote_write to Prometheus to allow Tempo
to add span metrics (allows us see which services are producing lots of
spans), and metrics are used to generate service graphs.

Ideally this chart should be deployed and managed by ArgoCD, but by having
everything in this Terraform file makes it easier to remove if we decide
against using Tempo.