Fix overquota on sg-rules-per-sg on origin LB SG.
Fix having too may security group rules on the www-origin load balancer
security group by using a single rule for port 80 and 443. Unfortunately
EC2 security groups are too simplistic to represent a set of ports and
can only represent a range, so use a range.

Mildly hacky but no practical implications because the LB only listens
on 80 and 443. So it's moot that we're theoretically not blocking
traffic to those ports with SG rules, because any such traffic would be
rejected at L3 by the LB anyway.

The risk of keeping separate rules is significant OTOH, since Terraform
doesn't handle the over-quota error path at all well (in fact it can
easily break everything and then leave the user to pick up the pieces by
hand).

(I've also increased the quotas anyway, but it still makes sense to
minimise the fanout on sg-rules here because defence in depth.)